Create releasefiles_rt index and SQL table structure / add to create and populate scripts

+++ misc/sphinxsearch/create_se_tables.php
@@ -5,8 +5,10 @@ if (nZEDb_RELEASE_SEARCH_TYPE != ReleaseSearch::SPHINX) {
 	exit('Error, nZEDb_RELEASE_SEARCH_TYPE in www/settings.php must be set to SPHINX!' . PHP_EOL);
 }
 
-if (!isset($argv[1]) || !isset($argv[2]) || !is_numeric($argv[2])) {
-	exit('Argument 1 must the hostname or IP to the Sphinx searchd server, Argument 2 must be the port to the Sphinx searchd server.' . PHP_EOL);
+if (!isset($argv[1]) || !isset($argv[2]) || !is_numeric($argv[2]) || !isset($argv[3]) || !is_bool((bool)$argv[3])) {
+	exit(
+		'Argument 1 must the hostname or IP to the Sphinx searchd server, Argument 2 must be the port to the Sphinx searchd server.' . PHP_EOL .
+		'The third and final argument is true/false.  True will drop tables if they exist and false will not.' . PHP_EOL);
 }
 
 $pdo = new nzedb\db\DB();
@@ -30,8 +32,24 @@ sprintf(
 $sphinxConnection
 );
 
+$tables['releasefiles_se'] =
+sprintf(
+"CREATE TABLE releasefiles_se
+(
+	id          BIGINT UNSIGNED NOT NULL,
+	weight      INTEGER NOT NULL,
+	query       VARCHAR(3072) NOT NULL,
+	releaseid   BIGINT UNSIGNED NOT NULL,
+	name        VARCHAR(255) NOT NULL DEFAULT '',
+	INDEX(query)
+) ENGINE=SPHINX CONNECTION=\"sphinx://%s/releasefiles_rt\"",
+$sphinxConnection
+);
+
 foreach ($tables as $table => $query) {
-	$pdo->queryExec(sprintf('DROP TABLE IF EXISTS %s', $table));
+	if ($argv[3] == true) {
+		$pdo->queryExec(sprintf('DROP TABLE IF EXISTS %s', $table));
+	}
 	$pdo->queryExec($query);
 }
